Произведен рефакторинг метода convert
